+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Changelog
-- sync upstream `1.21-0.32.60`
-- update NeoForge version `21.1.226`
+- support [BetterTab](https://modrinth.com/mod/better-tab)'s server data